requirements script updated as per CentOS9 Stream9

+++ requirements.sh
@@ -20,22 +20,24 @@
 ## Author: Mark Riddoch, Massimiliano Pinto
 ##
 
-fledge_location=`pwd`
-os_name=`(grep -o '^NAME=.*' /etc/os-release | cut -f2 -d\" | sed 's/"//g')`
-os_version=`(grep -o '^VERSION_ID=.*' /etc/os-release | cut -f2 -d\" | sed 's/"//g')`
+fledge_location=$(pwd)
+os_name=$(grep -o '^NAME=.*' /etc/os-release | cut -f2 -d\" | sed 's/"//g')
+os_version=$(grep -o '^VERSION_ID=.*' /etc/os-release | cut -f2 -d\" | sed 's/"//g')
 echo "Platform is ${os_name}, Version: ${os_version}"
 
-if [[ ( $os_name == *"Red Hat"* || $os_name == *"CentOS"* ) &&  $os_version == *"7"* ]]; then
-	echo Installing development tools 7 components
-	sudo yum install -y yum-utils
-	sudo yum-config-manager --enable rhel-server-rhscl-7-rpms
-	sudo yum install -y devtoolset-7
-	echo Installing boost components
-	sudo yum install -y boost-filesystem
-	sudo yum install -y boost-program-options
-	source scl_source enable devtoolset-7
-	export CC=/opt/rh/devtoolset-7/root/usr/bin/gcc
-	export CXX=/opt/rh/devtoolset-7/root/usr/bin/g++
+if [[ ( ${os_name} == *"Red Hat"* || ${os_name} == *"CentOS"* ) ]]; then
+    echo "Installing boost components..."
+	sudo yum install -y boost-filesystem boost-program-options
+
+	if [[ ${os_version} == *"7"* ]]; then
+        echo "Installing development tools 7 components..."
+        sudo yum install -y yum-utils
+        sudo yum-config-manager --enable rhel-server-rhscl-7-rpms
+        sudo yum install -y devtoolset-7
+        source scl_source enable devtoolset-7
+        export CC=/opt/rh/devtoolset-7/root/usr/bin/gcc
+        export CXX=/opt/rh/devtoolset-7/root/usr/bin/g++
+    fi
 elif apt --version 2>/dev/null; then
 	echo Installing boost components
 	sudo apt install -y libboost-filesystem-dev
@@ -45,19 +47,15 @@ else
 	echo "Requirements cannot be automatically installed, please refer README.rst to install requirements manually"
 fi
 
-if [ $# -eq 1 ]; then
+if [[ $# -eq 1 ]]; then
 	directory=$1
-	if [ ! -d $directory ]; then
-		mkdir -p $directory
-	fi
+	if [[ ! -d $directory ]]; then mkdir -p $directory; fi
 else
 	directory=~
 fi
 
 cd $directory
-if [ -d freeopcua ]; then
-	rm -rf freeopcua
-fi
+if [[ -d freeopcua ]]; then rm -rf freeopcua; fi
 
 echo Fetching Free OPCUA library
 git clone https://github.com/dianomic/freeopcua.git
@@ -76,5 +74,5 @@ cd build
 cmake ..
 make
 cd ..
-echo Set the environment variable FREEOPCUA to `pwd`
-echo export FREEOPCUA=`pwd`
+echo Set the environment variable FREEOPCUA to $(pwd)
+echo export FREEOPCUA=$(pwd)
